refactor(banners): change readme location
