+++ packages/snap-preact-components/src/components/Atoms/Merchandising/Banner/Banner.stories.tsx
@@ -5,7 +5,7 @@ import { ArgsTable, PRIMARY_STORY } from '@storybook/addon-docs/blocks';
 import { Banner, BannerProps } from './Banner';
 import { componentArgs } from '../../../../utilities';
 import { Snapify } from '../../../../utilities/snapify';
-import Readme from '../readme.md';
+import Readme from './readme.md';
 import { SearchController } from '@searchspring/snap-controller';
 
 export default {
